Revise dtype.

+++ mmaction/datasets/pipelines/loading.py
@@ -89,7 +89,7 @@ class SampleFrames(object):
             ratio = (num_frames - ori_clip_len + 1.0) / self.num_clips
             clip_offsets = np.around(np.arange(self.num_clips) * ratio)
         else:
-            clip_offsets = np.zeros((self.num_clips, ))
+            clip_offsets = np.zeros((self.num_clips, ), dtype=np.int32)
 
         return clip_offsets
 
@@ -115,7 +115,7 @@ class SampleFrames(object):
             if self.twice_sample:
                 clip_offsets = np.concatenate([clip_offsets, base_offsets])
         else:
-            clip_offsets = np.zeros((self.num_clips, ))
+            clip_offsets = np.zeros((self.num_clips, ), dtype=np.int32)
         return clip_offsets
 
     def _sample_clips(self, num_frames):
@@ -343,7 +343,7 @@ class SampleProposalFrames(SampleFrames):
             offsets = base_offsets + np.random.randint(
                 avg_interval, size=num_segments)
         else:
-            offsets = np.zeros((num_segments, ))
+            offsets = np.zeros((num_segments, ), dtype=np.int32)
 
         return offsets
 
@@ -367,7 +367,7 @@ class SampleProposalFrames(SampleFrames):
             base_offsets = np.arange(num_segments) * avg_interval
             offsets = (base_offsets + avg_interval / 2.0).astype(np.int32)
         else:
-            offsets = np.zeros((num_segments, ))
+            offsets = np.zeros((num_segments, ), dtype=np.int32)
 
         return offsets
